refactor: Move constants file
