--- public/collision/index.js
@@ -1,3 +1,0 @@
-import {Vector3} from "../../src/math/index.js";
-
-export const PLAYER_COLLISION_HULL = new Vector3(33, 73, 33);
