+++ public/collision/scenes/swept_aabb.js
@@ -6,7 +6,7 @@ import {Material} from "../../../src/Material/Material.js";
 import {Matrix3, PI, Vector2, Vector3} from "../../../src/math/index.js";
 import {Mesh} from "../../hl2/Mesh.js";
 import {Scene} from "../Scene.js";
-import {PLAYER_COLLISION_HULL} from "../index.js";
+import {ENTITY_HEIGHT_STAND, PLAYER_COLLISION_HULL} from "../../index.js";
 import {FIELD_OF_VIEW, SENSITIVITY} from "../main.js";
 
 /**
@@ -42,7 +42,7 @@ export async function createScene(imageBitmaps) {
 		}),
 		"playerHitbox",
 	);
-	playerHitbox.setPosition(new Vector3(0, PLAYER_COLLISION_HULL[1], -128));
+	playerHitbox.setPosition(new Vector3(0, ENTITY_HEIGHT_STAND, -128));
 	playerHitbox.buildHitbox();
 
 	meshes.push(playerHitbox);
@@ -66,10 +66,8 @@ export async function createScene(imageBitmaps) {
 export function createCamera(aspectRatio) {
 	const camera = new Camera();
 
-	camera.setPosition(new Vector3(0, PLAYER_COLLISION_HULL[1], -128));
+	camera.setPosition(new Vector3(0, ENTITY_HEIGHT_STAND, -128));
 	camera.target = new Vector3(camera.getPosition());
-	camera.setRotation(new Vector3(-PI / 6, 0, 0));
-	// camera.setDistance(new Vector3(0, 0, -64));
 
 	camera.fieldOfView = FIELD_OF_VIEW;
 	camera.aspectRatio = aspectRatio;
