+++ public/index.js
@@ -1,0 +1,5 @@
+import {Vector3} from "../src/math/index.js";
+
+export const ENTITY_HEIGHT_STAND = 64;
+export const FIELD_OF_VIEW = 75;
+export const PLAYER_COLLISION_HULL = new Vector3(33, 73, 33);
