+++ public/main.js
@@ -1,5 +1,5 @@
 import {setPolyfills} from "../src/polyfill.js";
-import main from "./collision/main.js";
+import main from "./visibility/main.js";
 
 setPolyfills();
 main();
